+++ packages/react-bindings/src/hooks/callStateHooks.ts
@@ -375,6 +375,7 @@ export const useCameraState = ({
   const { state } = camera;
   const direction = useObservableValue(state.direction$);
   const mediaStream = useObservableValue(state.mediaStream$);
+  const rootMediaStream = useObservableValue(state.rootMediaStream$);
   const selectedDevice = useObservableValue(state.selectedDevice$);
   const { getDevices } = useLazyDeviceList(camera);
   const hasBrowserPermission = useObservableValue(state.hasBrowserPermission$);
@@ -386,6 +387,7 @@ export const useCameraState = ({
     camera,
     direction,
     mediaStream,
+    rootMediaStream,
     get devices() {
       return getDevices();
     },