+++ packages/react-native-sdk/src/components/Participant/ParticipantView/ParticipantLabel.tsx
@@ -1,5 +1,11 @@
 import React, { useMemo } from 'react';
-import { Pressable, StyleSheet, Text, View } from 'react-native';
+import {
+  Pressable,
+  StyleSheet,
+  Text,
+  View,
+  ActivityIndicator,
+} from 'react-native';
 import {
   BadNetwork,
   MicOff,
@@ -7,7 +13,11 @@ import {
   ScreenShareIndicator,
   VideoSlash,
 } from '../../../icons';
-import { useCall, useI18n } from '@stream-io/video-react-bindings';
+import {
+  useCall,
+  useI18n,
+  useIsAudioConnecting,
+} from '@stream-io/video-react-bindings';
 import { ComponentTestIds } from '../../../constants/TestIds';
 import { type ParticipantViewProps } from './ParticipantView';
 import { Z_INDEX } from '../../../constants';
@@ -55,6 +65,7 @@ export const ParticipantLabel = ({
   const isAudioMuted = !hasAudio(participant);
   const isVideoMuted = !hasVideo(participant);
   const isTrackPaused = trackType && hasPausedTrack(participant, trackType);
+  const isAudioConnecting = useIsAudioConnecting(participant);
 
   if (trackType === 'screenShareTrack') {
     const screenShareText = isLocalParticipant
@@ -104,6 +115,13 @@ export const ParticipantLabel = ({
       ]}
     >
       <View style={styles.wrapper}>
+        {isAudioConnecting && (
+          <ActivityIndicator
+            size="small"
+            color={colors.iconPrimary}
+            style={styles.audioConnectingIndicator}
+          />
+        )}
         <Text style={[styles.userNameLabel, userNameLabel]} numberOfLines={1}>
           {participantLabel}
         </Text>
@@ -173,6 +191,10 @@ const useStyles = () => {
           fontSize: 13,
           fontWeight: '400',
           color: theme.colors.textPrimary,
+        },
+        audioConnectingIndicator: {
+          marginRight: theme.variants.spacingSizes.sm,
+          justifyContent: 'center',
         },
         screenShareIconContainer: {
           marginRight: theme.variants.spacingSizes.sm,